+++ HLSpriteKit/HLGestureTarget.h
@@ -19,18 +19,18 @@
 #import <UIKit/UIKit.h>
 
 /**
- Returns true if the passed gesture recognizers are of the same type and are configured
- in an equivalent way (dependent on class).
+ Returns true if the passed gesture recognizers are of the same type and are configured in
+ an equivalent way (dependent on class).
 
  For example, if the two passed gesture recognizers are both `UITapGestureRecognizers`
  configured with the same number of required taps and touches, then this method will
  return `YES`.
 
- Use case: Gesture targets return a list of gesture recognizers to which they might
- add themselves.  It is then the responsibility of the `UIGestureRecognizer` delegate
- (usually an `SKScene` or `UIViewController`) to add gesture recognizers to the view.
- But if the delegate already has an equivalent gesture recognizer added, then there's
- no need to add another.  This method can be used to decide what counts as "equivalent".
+ Use case: Gesture targets return a list of gesture recognizers to which they might add
+ themselves.  It is then the responsibility of the `UIGestureRecognizer` delegate (usually
+ an `SKScene` or `UIViewController`) to add gesture recognizers to the view.  But if the
+ delegate already has an equivalent gesture recognizer added, then there's no need to add
+ another.  This method can be used to decide what counts as "equivalent".
 
  @bug Might be worth comparing and contrasting with `[UIGestureTarget isEqual:]`.
 */
@@ -39,7 +39,7 @@ BOOL HLGestureTarget_areEquivalentGestureRecognizers(UIGestureRecognizer *a, UIG
 /**
  A generic target for `UIGestureRecognizers`.
 
- ## Use Case
+ ## Usage
 
  A single delegate for a bunch of gesture recognizers creates and maintains the
  recognizers, but wants to forward the gesture to different targets based on where the
@@ -71,6 +71,46 @@ BOOL HLGestureTarget_areEquivalentGestureRecognizers(UIGestureRecognizer *a, UIG
         [myButton hlSetGestureTarget:[[HLTapGestureTarget alloc] initWithHandleGestureBlock:^{
             NSLog(@"tapped button");
         }]];
+
+ ## Ad Hoc Gesture Targets: Benefits and Dangers
+
+ Implementation (3), above, describes a usage pattern that allows ad hoc creation of
+ interactive nodes, without subclassing.
+
+ It can be nice to do this for a simple one-off popup: a message, perhaps, or an extremely
+ simple configuration dialog.
+
+ Say I want to create a popup message with a single Done button, in an ad hoc fashion.
+ Easy.  (I'm leaving out the sizing and positioning code so it doesn't distract from the
+ logic.)
+
+     SKSpriteNode *alertNode = [SKSpriteNode spriteNodeWithColor:[SKColor blueColor] size:CGSizeMake(200.0f, 180.0f)];
+
+     HLMultilineLabelNode *labelNode = [[HLMultilineLabelNode alloc] initWithFontNamed:@"Helvetica"];
+     labelNode.text = @"Something happened, and you should know about it.";
+
+     HLLabelButtonNode *doneButton = [[HLLabelButtonNode alloc] initWithColor:[SKColor blackColor] size:CGSizeMake(80.0f, 40.0f)];
+     doneButton.text = @"Done";
+
+     [alertNode addChild:labelNode];
+     [alertNode addChild:doneButton];
+
+     __weak SKSpriteNode *alertNodeWeak = alertNode;
+     [doneButton hlSetGestureTarget:[HLTapGestureTarget tapGestureTargetWithHandleGestureBlock:^(UIGestureRecognizer *gestureRecognizer){
+       [alertNodeWeak removeFromParent];
+     }]];
+     [self needSharedGestureRecognizersForNode:doneButton];
+
+     [self addChild:alertNode];
+
+ But already it's a bit of a pain setting up the tap gesture block with a weak reference
+ (in order to avoid a retain cycle on `alertNode`).  And it can quickly get much more
+ complicated when there is more to do in the block: for instance, reading user input
+ values from the dialog, or having separate "Okay" and "Cancel" buttons.
+
+ Anyway, the point is: `HLSpriteKit` allows ad hoc gesture targets, which is nice, but in
+ many instances subclassing (or at least delegation) makes things much simpler and easier
+ to understand.
 */
 @protocol HLGestureTarget <NSObject>
 
@@ -202,6 +242,36 @@ BOOL HLGestureTarget_areEquivalentGestureRecognizers(UIGestureRecognizer *a, UIG
 
 /**
  A block that will be executed when the gesture target is tapped.
+
+ Beware retain cycles when using the callback to invoke a method on the node that owns
+ this gesture target.  A common example is creating a node that dismisses itself when
+ tapped:
+
+     [myNode hlSetGestureTarget:[HLTapGestureTarget tapGestureTargetWithHandleGestureBlock:^(UIGestureRecognizer *gestureRecognizer){
+       [myNode removeFromParent];
+     }]];
+
+ In the example, `myNode` retains its gesture target, which has a block retaining
+ `myNode`.  This can be rewritten to retain the node weakly:
+
+     __weak SKNode *myNodeWeak = myNode;
+     [myNode hlSetGestureTarget:[HLTapGestureTarget tapGestureTargetWithHandleGestureBlock:^(UIGestureRecognizer *gestureRecognizer){
+       [myNodeWeak removeFromParent];
+     }]];
+
+ If the weak reference is mentioned more than once, then it might need to be made strong
+ again inside the block:
+
+     __weak SKNode *myNodeWeak = myNode;
+     [myNode hlSetGestureTarget:[HLTapGestureTarget tapGestureTargetWithHandleGestureBlock:^(UIGestureRecognizer *gestureRecognizer){
+       SKNode *myNodeStrongAgain = myNodeWeak;
+       if (myNodeStrongAgain.parent) {
+         [myNodeStrongAgain removeFromParent];
+       }
+     }]];
+
+ An easier way to avoid retain cycles is to use the delegate interface of
+ `HLTapGestureTarget`; see `delegate`.
 */
 @property (nonatomic, strong) void (^handleGestureBlock)(UIGestureRecognizer *);
 
@@ -222,104 +292,5 @@ BOOL HLGestureTarget_areEquivalentGestureRecognizers(UIGestureRecognizer *a, UIG
 - (void)tapGestureTarget:(HLTapGestureTarget *)tapGestureTarget didTap:(UIGestureRecognizer *)gestureRecognizer;
 
 @end
-
-/**
- noob: So in an HLScene I was working on, I had a method for creating a modal
- presentation consisting of a number of nodes which I created on-demand in the scene
- without using any custom node subclassing.  I used a bit SKSpriteNode in the background
- with an HLTapGestureTarget to handle taps, which would dismiss the modal presentation;
- also, another SKSpriteNode button (again with an HLTapGestureTarget) did the same
- thing, but with a special function (go to the next level of the game).  I did it
- without creating a dismissal callback in the scene, so then all nodes and behavior was
- configured once at setup, and then displayed, with no other state necessary in the
- scene.
-
- So...it's pretty messy, because the cleanup needs to be extremely careful about retain
- cycles.  Furthermore, it gets more complicated when there are multiple gesture targets
- on the modal presentation: For example, each of them needs to unregister themselves and
- *all* the others from the HLScene.  Nasty.
-
- Here were my thoughts (before I eventually went ahead and subclassed):
-
-  1) The buttons need to be aware of each other.  Perhaps like the "Okay" and "Cancel"
-     buttons of an alert, all actions should pass through the same callback.  In that
-     case, the owner has an FL_goalsDismissWithButtonIndex method, with stored
-     goalsOverlay state from this method, and set each of their handleGestureBlocks to
-     call it.  But that seems to be getting closer and closer to subclassing: The
-     buttons are acting together, with shared state, and so should be entirely
-     encapsulated together.
-
-  2) But really, the only reason the buttons need to be aware of each other is
-     (currently) because of unregistering: they both need to unregister both (when
-     dismissing the overlay).  Which reminds me that unregistering HLGestureTargets is a
-     pain in the ass in general, and according to current implementation not even
-     essential.  BUT.  Unregistering still makes sense for other possible future HLScene
-     implementations, and no matter what, unregistering is a nice option to have (even
-     just to clear userData) and so it philosophically makes sense to always do it.
-
-  3) Unregistering is especially a pain in the ass when the HLTapGestureTarget wants to
-     unregister its owner node: The node contains a reference to the handleGesture
-     block, but then we try to make the block contain a reference to the node.  To break
-     the retain cycle, we can make the node reference weak, but that's just one more
-     line of code in something that already feels unnecessary.  Can there be
-     functionality somewhere which automatically unregisters a node when the node is
-     deallocated?
-
-  4) And in fact the real problem is HLGestureTarget nodes that don't just want to
-     unregister but in fact want to delete themselves.  Very common: Create some kind of
-     dialog box, and add a single button which dismisses it.  So then the button removes
-     the dialog box from the node hierarchy, no other references exist, the parent is
-     deleted which deletes the children, the button is deleted, so the callback block
-     (being run) is deleted.  So (see notes in notes/objective-c.txt) we have to add TWO
-     lines of code, making a strong reference (at block execution time) of a weak
-     reference (at block copy time) of the dialog box.  What a pain.  HLGestureTarget
-     should make this easier for us somehow.  Could it retain a strong reference for us
-     right before invoking the block?
-
- For now: Consider it normal that, when building a node with multiple out-of-the-box
- HLGestureTargets, you have to set their handleGesture callbacks all at the same time at
- the bottom of the setup code, with full awareness of each other.
-
- The code evolved as far as this before I subclassed:
-
-     __weak HLLabelButtonNode *victoryButtonWeak = victoryButton;
-     __weak SKSpriteNode *dismissNodeWeak = dismissNode;
-     __weak HLScrollNode *goalsOverlayWeak = goalsOverlay;
-
-     if (victoryButton) {
-       [victoryButton hlSetGestureTarget:[[HLTapGestureTarget alloc] initWithHandleGestureBlock:^(UIGestureRecognizer *gestureRecognizer){
-         if (self->_tutorialState.tutorialActive) {
-           [self FL_tutorialRecognizedAction:FLTutorialActionGoalsDismissed withArguments:nil];
-         }
-         [self unregisterDescendant:victoryButtonWeak];
-         [self unregisterDescendant:dismissNodeWeak];
-         [self unregisterDescendant:goalsOverlayWeak];
-         self->_goalsState.clear();
-         // noob: Retain a strong reference to block owner when dismissing the modal node; nobody else
-         // is retaining the victoryButton, but we'd like to finish running this block before getting
-         // deallocated.  The weak reference is copied with the block at copy time; now this strong
-         // reference (though theoretically possibly nil) will exist until we're done the block.  It's
-         // not actually clear how necessary this is, because I don't usually see problems unless this
-         // block starts deleting a whole bunch of stuff (like if the didTapNext delegate method deletes
-         // the scene right away, as it is prone to do if it is not careful).
-         __unused HLLabelButtonNode *victoryButtonStrongAgain = victoryButtonWeak;
-         [self dismissModalNodeAnimation:HLScenePresentationAnimationNone];
-         id<FLTrackSceneDelegate> delegate = self.delegate;
-         if (delegate) {
-           // noob: So this is dangerous.  The delegate is probably going to delete this scene.
-           // We've got strong references to the scene copied with the block, so let's make sure
-           // the block is gone before we try to deallocate the scene.  Okay so wait that's a problem
-           // with all existing blocks that reference self, right?  Like, they should all have __weak
-           // references?  Unless SKNode explicitly releases children during its deallocation.
-           // Sooooo . . . that's something to test.  For now, there aren't crashes, and if there's
-           // a retain cycle I haven't noticed it yet.
-           [delegate performSelector:@selector(trackSceneDidTapNextLevelButton:) withObject:self];
-         }
-       }]];
-       [self registerDescendant:victoryButton withOptions:[NSSet setWithObject:HLSceneChildGestureTarget]];
-     }
-
-  ...etc...
-*/
 
 #endif